// frontend/src/openapi/models/Span.ts
export interface Span {
    /**
     *
     * @type {string}
     * @memberof Span
     */
    readonly spanId: string;
    /**
     *
     * @type {string}
     * @memberof Span
     */
    readonly traceId: string;
    /**
     *
     * @type {string}
     * @memberof Span
     */
    readonly parentSpanId: string | null;
    /**
     *
     * @type {boolean}
     * @memberof Span
     */
    readonly isRoot: boolean;
    /**
     *
     * @type {string}
     * @memberof Span
     */
    readonly project: string;
    /**
     *
     * @type {string}
     * @memberof Span
     */
    readonly spanType: string;
    /**
     *
     * @type {string}
     * @memberof Span
     */
    readonly operation: string;
    /**
     *
     * @type {string}
     * @memberof Span
     */
    readonly name: string;
    /**
     *
     * @type {number}
     * @memberof Span
     */
    readonly kind: number;
    /**
     *
     * @type {number}
     * @memberof Span
     */
    readonly startTimeNs: number;
    /**
     *
     * @type {number}
     * @memberof Span
     */
    readonly endTimeNs: number;
    /**
     *
     * @type {number}
     * @memberof Span
     */
    readonly durationNs: number;
    /**
     *
     * @type {number}
     * @memberof Span
     */
    readonly statusCode: number;
    /**
     *
     * @type {string}
     * @memberof Span
     */
    readonly statusMessage: string;
    /**
     *
     * @type {string}
     * @memberof Span
     */
    readonly serviceName: string;
    /**
     *
     * @type {any}
     * @memberof Span
     */
    readonly resourceAttrs: any | null;
    /**
     *
     * @type {string}
     * @memberof Span
     */
    readonly scopeName: string;
    /**
     *
     * @type {string}
     * @memberof Span
     */
    readonly scopeVersion: string;
    /**
     *
     * @type {any}
     * @memberof Span
     */
    readonly attributes: any | null;
    /**
     *
     * @type {any}
     * @memberof Span
     */
    readonly events: any | null;
    /**
     *
     * @type {any}
     * @memberof Span
     */
    readonly links: any | null;
    /**
     *
     * @type {string}
     * @memberof Span
     */
    readonly capability: string | null;
    /**
     *
     * @type {string}
     * @memberof Span
     */
    readonly conversation: string | null;
    /**
     *
     * @type {any}
     * @memberof Span
     */
    readonly feedbackScore: any | null;
    /**
     *
     * @type {Date}
     * @memberof Span
     */
    readonly receivedAt: Date;
}