// frontend/src/openapi/models/Verdict.ts
export interface Verdict {
    /**
     *
     * @type {string}
     * @memberof Verdict
     */
    readonly id: string;
    /**
     *
     * @type {string}
     * @memberof Verdict
     */
    readonly evaluator: string | null;
    /**
     *
     * @type {string}
     * @memberof Verdict
     */
    readonly evaluatorName: string;
    /**
     *
     * @type {string}
     * @memberof Verdict
     */
    readonly displayName: string;
    /**
     *
     * @type {TargetKindEnum}
     * @memberof Verdict
     */
    readonly targetKind: TargetKindEnum;
    /**
     *
     * @type {string}
     * @memberof Verdict
     */
    readonly targetId: string;
    /**
     *
     * @type {string}
     * @memberof Verdict
     */
    readonly label: string;
    /**
     *
     * @type {number}
     * @memberof Verdict
     */
    readonly score: number | null;
    /**
     *
     * @type {boolean}
     * @memberof Verdict
     */
    readonly passed: boolean | null;
    /**
     *
     * @type {VerdictOutcomeEnum}
     * @memberof Verdict
     */
    readonly outcome: VerdictOutcomeEnum;
    /**
     *
     * @type {string}
     * @memberof Verdict
     */
    readonly explanation: string;
    /**
     *
     * @type {any}
     * @memberof Verdict
     */
    readonly unmet: any | null;
    /**
     *
     * @type {string}
     * @memberof Verdict
     */
    readonly scope: string;
    /**
     *
     * @type {string}
     * @memberof Verdict
     */
    readonly grain: string;
    /**
     *
     * @type {boolean}
     * @memberof Verdict
     */
    readonly gate: boolean;
    /**
     *
     * @type {Array<any>}
     * @memberof Verdict
     */
    readonly subScores: Array<any>;
    /**
     *
     * @type {AnnotatorKindEnum}
     * @memberof Verdict
     */
    readonly annotatorKind: AnnotatorKindEnum;
    /**
     *
     * @type {string}
     * @memberof Verdict
     */
    readonly identifier: string;
    /**
     *
     * @type {string}
     * @memberof Verdict
     */
    readonly judgeTraceId: string;
    /**
     *
     * @type {number}
     * @memberof Verdict
     */
    readonly cost: number | null;
    /**
     *
     * @type {number}
     * @memberof Verdict
     */
    readonly latencyMs: number | null;
    /**
     *
     * @type {Date}
     * @memberof Verdict
     */
    readonly createdAt: Date;
    /**
     *
     * @type {Date}
     * @memberof Verdict
     */
    readonly updatedAt: Date;
}